First Deliverable
For this First Deriverable we have:
   *The creation of the footer, for this we have the folder components and the footer component inside.
We installed the router to save time in the future. Under index.js we will find the first part and the component PageRouter will contain the rest of the routes.
The Footer component was created too as a component.
Under the Home component we will have the ContentHome that will host the three main components of this deliverable:
    The Carrusel
    The categories
    Finally the featured products
The carrusel called slider as a component and the other two as Categories and FeaturedProducts.

For the future should be better to create tests and have the cart icon as individual component since it uses the same module as the store logo.